finish qsort skeleton & add multithreading
